+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 ## v0.0.46 [unreleased]
 
 - version 0.0.46
+- `ryo3-reqwest`
+  - `ResponseStream`
+    - Added `__repr__` method
+    - Added `async def take(self, n: int=1): ...` method returns n chunks as a
+      list
+    - Added `async def collect(self: join = False) -> ...:` method that collects
+      the stream into a single `ry.Bytes` object if `join=True` or a list of
+      `ry.Bytes` objects if `join=False`
+    - Added `async def take(self, n: int=1): ...` which returns n chunks as a list
 - `ryo3-glob`
   - add `dtype` kwarg that takes either `dtype=str | ry.FsPath | pathlib.Path`
     as type of obj yielded by the iterable; something about this feels really